refactor: Create numtools module, move relevant functions
- Created numtools as a generic numerical function module, separate from
  common
- common will continue to contain more generic utilty functions
