+++ vtkpytools/__init__.py
@@ -1,8 +1,9 @@
 from .barfiletools import *
 from .gridtools3d import *
 from .gridtools2d import *
-from .common import (getGeometricSeries, unstructuredToPoly, orderPolyDataLine,
-                     vCutter, Profile, globFile, rotateTensor, makeRotationTensor,
-                     symmetric2FullTensor, full2SymmetricTensor, calcStrainRate)
+from .common import (unstructuredToPoly, orderPolyDataLine,
+                     vCutter, Profile, globFile,)
+from .numtools import (getGeometricSeries, rotateTensor, makeRotationTensor,
+                       symmetric2FullTensor, full2SymmetricTensor, calcStrainRate)
 from .bl import (integratedVortBLThickness, sampleAlongVectors, delta_vortInt, delta_velInt, delta_percent)
